update hook for stramer
